Move tests over and refactor initialized datastore
